refactor mutation
