+++ src/main/java/jp/kusumotolab/kgenprog/ga/mutation/SimpleMutation.java
@@ -1,27 +1,15 @@
 package jp.kusumotolab.kgenprog.ga.mutation;
 
-import java.util.ArrayList;
-import java.util.List;
 import java.util.Random;
-import java.util.function.Function;
 import org.eclipse.jdt.core.dom.ASTNode;
-import jp.kusumotolab.kgenprog.fl.Suspiciousness;
-import jp.kusumotolab.kgenprog.ga.Roulette;
 import jp.kusumotolab.kgenprog.ga.mutation.Scope.Type;
 import jp.kusumotolab.kgenprog.ga.mutation.selection.CandidateSelection;
-import jp.kusumotolab.kgenprog.ga.validation.Fitness;
-import jp.kusumotolab.kgenprog.ga.variant.Base;
-import jp.kusumotolab.kgenprog.ga.variant.Gene;
-import jp.kusumotolab.kgenprog.ga.variant.HistoricalElement;
-import jp.kusumotolab.kgenprog.ga.variant.MutationHistoricalElement;
-import jp.kusumotolab.kgenprog.ga.variant.Variant;
-import jp.kusumotolab.kgenprog.ga.variant.VariantStore;
 import jp.kusumotolab.kgenprog.project.ASTLocation;
 import jp.kusumotolab.kgenprog.project.FullyQualifiedName;
 import jp.kusumotolab.kgenprog.project.NoneOperation;
 import jp.kusumotolab.kgenprog.project.Operation;
-import jp.kusumotolab.kgenprog.project.jdt.InsertAfterOperation;
 import jp.kusumotolab.kgenprog.project.jdt.DeleteOperation;
+import jp.kusumotolab.kgenprog.project.jdt.InsertAfterOperation;
 import jp.kusumotolab.kgenprog.project.jdt.InsertBeforeOperation;
 import jp.kusumotolab.kgenprog.project.jdt.ReplaceOperation;
 
@@ -33,7 +21,6 @@ import jp.kusumotolab.kgenprog.project.jdt.ReplaceOperation;
 public class SimpleMutation extends Mutation {
 
   protected final Type type;
-  private final boolean needHistoricalElement;
 
   /**
    * コンストラクタ
@@ -47,60 +34,8 @@ public class SimpleMutation extends Mutation {
   public SimpleMutation(final int mutationGeneratingCount, final Random random,
       final CandidateSelection candidateSelection, final Type type,
       final boolean needHistoricalElement) {
-    super(mutationGeneratingCount, random, candidateSelection);
+    super(mutationGeneratingCount, random, candidateSelection, needHistoricalElement);
     this.type = type;
-    this.needHistoricalElement = needHistoricalElement;
-  }
-
-  /**
-   * 乱数に基づいて選択された Variant に対して変異処理をする Variant の選択は fitness の逆数に基づいて行われる 変異処理された Variant を
-   * mutationGeneratingCount 分だけ返す
-   *
-   * @param variantStore Variant の情報を格納するオブジェクト
-   * @return 変異された Gene を持った Variant のリスト
-   */
-  @Override
-  public List<Variant> exec(final VariantStore variantStore) {
-
-    final List<Variant> generatedVariants = new ArrayList<>();
-
-    final List<Variant> currentVariants = variantStore.getCurrentVariants();
-
-    final Roulette<Variant> variantRoulette = new Roulette<>(currentVariants, e -> {
-      final Fitness fitness = e.getFitness();
-      final double value = fitness.getValue();
-      return Double.isNaN(value) ? 0 : value;
-    }, random);
-
-    for (int i = 0; i < mutationGeneratingCount; i++) {
-      final Variant variant = variantRoulette.exec();
-      final List<Suspiciousness> suspiciousnesses = variant.getSuspiciousnesses();
-      final Function<Suspiciousness, Double> weightFunction = susp -> Math.pow(susp.getValue(), 2);
-
-      if (suspiciousnesses.isEmpty()) {
-        continue;
-      }
-      final Roulette<Suspiciousness> roulette =
-          new Roulette<>(suspiciousnesses, weightFunction, random);
-
-      final Suspiciousness suspiciousness = roulette.exec();
-      final Base base = makeBase(suspiciousness);
-      final Gene gene = makeGene(variant.getGene(), base);
-      final HistoricalElement element;
-      if (needHistoricalElement) {
-        element = new MutationHistoricalElement(variant, base);
-      } else {
-        element = null;
-      }
-      generatedVariants.add(variantStore.createVariant(gene, element));
-    }
-
-    return generatedVariants;
-  }
-
-  protected Base makeBase(final Suspiciousness suspiciousness) {
-    final ASTLocation location = suspiciousness.getLocation();
-    return new Base(location, makeOperation(location));
   }
 
   protected Operation makeOperation(final ASTLocation location) {
@@ -126,11 +61,5 @@ public class SimpleMutation extends Mutation {
     final Scope scope = new Scope(type, fqn);
     final Query query = new Query(scope);
     return candidateSelection.exec(query);
-  }
-
-  protected Gene makeGene(final Gene parent, final Base base) {
-    final List<Base> bases = new ArrayList<>(parent.getBases());
-    bases.add(base);
-    return new Gene(bases);
   }
 }
